API/requirements.txt
fastapi
uvicorn
pydantic
mecab-python3
unidic-lite